User setting to set how parameters are inserted
Autocomplete can insert parameters as either
None: no parameters (just commas)
Type: Inserts the types of the parameters
Name: Inserts the names of the parameters
Both: Inserts <type_name>

User can select which setting in the Atom preferences for the package.  Defaults to Type, which was the most common before this patch.